requirements.txt
openai
requests
python-telegram-bot==13.15